!leave and !endgame fixes
The !leave command did not work properly for me. When a player left during their turn the game would still freeze up. If they joined back and left again it would let the game continue. This is fixed now.

When a player leaves on their turn it will properly remove the player and switch the turn to the next player and display the "TELL USER TO GO" information.

Also when playing with the AI I noticed it did not leave the channel after using the !endgame command. Now it will.